Fixes around extra whitespace matching.
First, removed the matchdelete on buffer leave. It seems to not run exactly as
I think it will and can end up showing error messages at times. The easiest
repro seems to be:
1) Open a file
2) :tabe to a new file
3) Close the new file
4) Enter insert mode on the original file

I'm sure there is a way to write it correctly, but I'm not spending the time on
it, and removing the matchdelete appears to still work as expected around
insert mode.

Also, added the highlight for ExtraWhitespace to an autocmd to run after syntax
highlighting is changed. Otherwise, a colorscheme command will clear the
highlight definition.